refactor: move aphrodite to metrics
